Building Api with Lumen Authorization and Relationship

+++ app/Http/Controllers/BoardController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Board;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class BoardController extends Controller
 {
@@ -14,22 +15,26 @@ class BoardController extends Controller
      */
     public function __construct()
     {
-        //
+        $this->middleware('auth');
     }
 
     public function index(){
-      return Board::all();
+      return Auth::user()->boards;
     }
 
     public function show($id){
       $board = Board::findOrFail($id);
+
+      if (Auth::user()->id !== $board->user_id) {
+        return response()->json(['status'=>'error', 'message'=>'unauthorized'], 401);
+      }
+
       return $board;
     }
 
     public function store(Request $request){
-      Board::create([
-        'name' => $request->name,
-        'user_id' => $request->user_id
+      Auth::user()->boards()->create([
+        'name' => $request->name
       ]);
       return response()->json(['message'=>'success'], 200);
     }
@@ -37,7 +42,12 @@ class BoardController extends Controller
     public function update(Request $request, $id){
       $board = Board::findOrFail($id);
 
-      if ($board->update($request->all())) {
+      if (Auth::user()->id !== $board->user_id) {
+        return response()->json(['status'=>'error', 'message'=>'unauthorized'], 401);
+      }
+
+      $board->name = $request->name;
+      if ($board->save()) {
         return response()->json(['message'=>'success', 'board'=>$board], 200);
       }
 
@@ -45,6 +55,12 @@ class BoardController extends Controller
     }
 
     public function destroy($id){
+      $board = Board::findOrFail($id);
+
+      if (Auth::user()->id !== $board->user_id) {
+        return response()->json(['status'=>'error', 'message'=>'unauthorized'], 401);
+      }
+
       if (Board::destroy($id)) {
         return response()->json(['status'=>'success', 'message'=>'Board Deleted Successfully']);
       }
